requirements.txt
fastapi
uvicorn
boto3
python-dotenv
Pillow
python-multipart
celery[redis]
pika



